Create build operation in task executer
